camply-backend/requirements.txt
google-adk
asyncpg
asyncio
python-dotenv
sqlalchemy
psycopg2-binary
httpx
beautifulsoup4
lxml
requests
pydantic
fastapi
uvicorn
python-multipart
PyMuPDF>=1.24.0
pdfplumber>=0.10.0
PyMuPDF==1.23.14
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
python-dateutil
supabase>=2.15.0 